proxy/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.1
fakeredis>=2.20.0
clickhouse-connect>=0.7.0
structlog>=24.0.0
pydantic>=2.7.0
pydantic-settings>=2.4.0
psycopg2-binary>=2.9.9
stripe>=9.9.0
litellm>=1.40.0
